refactor: move titleFromPath to layout, document hideOnMobile in AGENTS.md
- NavBar becomes fully presentational: accepts optional pageTitle prop,
  no longer contains route knowledge (titleFromPath removed)
- +layout.svelte derives pageTitle from URL pathname and passes to NavBar
- AGENTS.md: document hideOnMobile column option in DataTable props table
